[electrum] Bugfix: close the "Check your trezor" popup after entering a passphrase
Summary: The clear signal that kills the popup is emitted but the flow don't wait for completion, causing a race when another flow is started. This causes the popup to remain on screen and raise an exception upon closure.

Test Plan: Unlock a trezor HW and check the popup is closed as expected.

Reviewers: #bitcoin_abc, bytesofman, PiRK

Reviewed By: #bitcoin_abc, bytesofman, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17547