[electrum] enable the Build Avalanche Proof context menu action when possible for Trezor wallets
Summary:
Some Trezor wallets with custom firmware can now sign stakes when a custom version of the trezor library is installed/packaged.
The Proof editor was already enabled when accessed via the Tools menu. Enable it as well when accessed via the coins tab with eligible coins selected.

Test Plan: With a trezor device with a custom firmware, select a coin an click Build Avalanche Proof in the context menu. Add a master private key and build the proof.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17787