[electrum] make it possible to spend tokens by manually selecting the coins to spend
Summary:
Instead of excluding all token utxos from the selected coins for  the "Spend" action in the context menu, just add an explicit warning dialog if any selected coin contains tokens . The user gets two chances to not spend the tokens: by clicking Cancel in the warning dialog, or Clear in the Spend tab if he accidentaly clicked on "I understand the risks"

This does not change the automatic coin selection behavior, tokens are still excluded from regular transactions unless they have been manually selected in the coins tab.

Depends on D17808

Test Plan: Open a wallet that has a mix of regular utxos, ALP tokens, SLP tokens. Try various coin selection and check that you get the warning whenever a token is selected. By dismissing the warning, check that the coins are part of the transaction (Preview button in the Send tab)

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17809