[electrum] include metadata for slip10 for the windows and osx releases
Summary: Similar to what was done in D17785, we need to make sure the slip10 library is packaged with its metadata.

Test Plan:
Build the windows release and test a trezor device. The error fixed by this diff would prevent connecting to the device, so it is enough to just successfully restore or open a Trezor  wallet.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D17819