[electrum] Catch a potential LibUSB error during session close when jumping to bootloader
Summary: The issue exists some machines/OS/devices (still unclear) that libusb raises an error when the session is closed during a reboot to bootloader. Since this does not affect the upgrade process we can safely catch and ignore this error as a workaround.

Test Plan: Tested on a failing machine/os/device combination.

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Subscribers: PiRK

Differential Revision: https://reviews.bitcoinabc.org/D17844