[electrum] remove shortcut reference to InvoiceStore in MainWindow
Summary: The wallet owns the InvoiceStore, and all the widgets that need to  access it have a reference to the wallet, so it is not necessary to use the `main_window.invoices` indirection

Test Plan: In the receive tab type an amount and click save (this adds an invoice to the store and shows the invoice list. Delete the invoice via the right click context menu in the invoice list.

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18329