[electrum] rename Stake.utxo to Stake.outpoint
Summary:
In most contexts utxo and outpoint are interchangeable words. But in Electrum ABC, utxo is usually associated with the complicated dict packed with metadata returned by `wallet.get_utxos()`. For avalanche stakes, this attribute is just an `Outpoint` object with a txid and output index.

Depends on D18770

Test Plan:
`python test_runner.py`

Try the avalanche proof editor

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18771