[lint] bump black to >=25
Summary:
This is the yearly bumping for the black linter. Note that v25.x  still supports python 3.9
See D15374 for the  previous bump.

But this time it is probably the last bump, as this is an intermediate step before replacing `black` with `ruff`. This helps making the `ruff` diff smaller by fixing some differences in advance.

Test Plan:
```
arc liberate
arc lint --everything
```

`ninja all check-all`

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Subscribers: Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18903