[lint] add a blank line after module level docstring in python files
Summary:
This is extracted out of D18904.

I used the search and replace tool in my IDE with the following regex:
`^("""(?:\\[\s\S]|[^"\\]|"(?!"")|""(?!"))*""")\r?\n(\w)`

Replaced by
`$1\n\n$2`

A few false positives had to be filtered out, so this is not really a scripted diff.

Depends on D18903

Test Plan:
Trivial self-review, and a run of all python tests just to be sure.
`ninja all check-all check-electrum check-electrum-functional``

Reviewers: #bitcoin_abc, Fabien

Reviewed By: #bitcoin_abc, Fabien

Differential Revision: https://reviews.bitcoinabc.org/D18912